backend/requirements.txt
bcrypt
cryptography
passlib
fastapi
uvicorn
python-jose
SQLAlchemy
python-multipart